requirements.txt
# Web Framework
Flask==2.3.2
flask-cors==4.0.0
flask-limiter==2.7.0
fastapi==0.104.1
uvicorn==0.27.0

# Document Processing
python-docx==0.8.11
docx2txt==0.8
pypdf==3.15.1
pdfplumber==0.10.2
pypdfium2==4.10.0

# Vector Store and Embedding
elasticsearch==8.11.0
langchain==0.1.0
langchain-community==0.0.13
langchain-elasticsearch==0.0.2
langchain-experimental==0.0.52
langchain-openai==0.0.5

# AI Services
openai==1.1.9
bert-extractive-summarizer==0.10.1
torch==2.1.1

# Database
pymongo==4.6.1
mysql-connector-python==8.2.0
sqlalchemy==2.0.23

# Visualization
pyvis==0.3.2

# Authentication
PyJWT==2.8.0
pydantic==2.4.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2